Giving up on merging files for now
Will just make individual castras. The issue with the current approach
for the merge was that it was appending the new objects in its own array
instead of  adding it to the existing json array. A bit hard to explain
but create an empty json file, add the contents of a subreddit with a
few entries, try to merge with another sub with few entries using the
glob method. Then do a json prettify and see how the structure is wrong.